Progressed with groups. Refactored roles and permissions. Added slug field, details in the migration

+++ app/Http/Controllers/Api/UserController.php
@@ -7,7 +7,6 @@ use App\Notifications\RegisterUser;
 use App\User;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
-use Illuminate\Http\Resources\Json\Resource;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Str;
@@ -38,7 +37,7 @@ class UserController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
@@ -54,16 +53,20 @@ class UserController extends Controller
         }
 
         $request->validate([
-           'name'=>'required',
-           'email'=> ['required', 'string', 'email', 'max:255', 'unique:tenant.users,email'],
-           'password'=> ['required', 'string', 'min:8', 'confirmed']
+            'name' => 'required',
+            'email' => ['required', 'string', 'email', 'max:255', 'unique:tenant.users,email'],
+            'password' => ['required', 'string', 'min:8', 'confirmed']
         ]);
 
-        $user= User::create([
-            'name'=> $request->name,
-            'email'=> $request->email,
-            'password'=> Hash::make($request->password)
+        $user = User::create([
+            'name' => ucwords($request->name),
+            'email' => $request->email,
+            'password' => Hash::make($request->password)
         ]);
+
+        // Assign the avatar;
+        $avatar_url = 'https://img.favpng.com/2/12/12/computer-icons-portable-network-graphics-user-profile-avatar-png-favpng-L1ihcbxsHbnBKBvjjfBMFGbb7.jpg';
+        $user->addMediaFromUrl($avatar_url)->toMediaCollection('avatar');
 
         if ($request->has('role')) {
             $user->assignRole($request->role['name']);
@@ -78,8 +81,8 @@ class UserController extends Controller
         $user->markEmailAsVerified();
 
         return response([
-            'message'=>'User Created',
-            'payload'=>UserResource::make($user)
+            'message' => 'User Created',
+            'payload' => UserResource::make($user)
         ], Response::HTTP_CREATED);
     }
 
@@ -87,20 +90,20 @@ class UserController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param \Illuminate\Http\Request $request
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, User $user)
     {
         $request->validate([
-            'name'=>'required',
-            'email'=>'required|email',
+            'name' => 'required',
+            'email' => ['required', 'string', 'email', 'max:255', 'unique:tenant.users,email,'.$user->id],
         ]);
 
         $user->update([
-            'name'=> $request->name,
-            'email'=> $request->email,
+            'name' => ucwords($request->name),
+            'email' => $request->email,
             'updated_at' => now()
         ]);
 
@@ -113,8 +116,8 @@ class UserController extends Controller
         }
 
         return response([
-            'message'=>'User Updated',
-            'payload'=> UserResource::make($user)
+            'message' => 'User Updated',
+            'payload' => UserResource::make($user)
         ], Response::HTTP_ACCEPTED);
 
     }
@@ -122,14 +125,20 @@ class UserController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\JsonResponse
      */
     public function destroy($id)
     {
-        User::destroy($id);
-        return response()->json([
-            'message' => 'User successfully removed'
-        ], 200);
+        if (Auth::user()->id != $id) {
+            User::destroy($id);
+            return response()->json([
+                'message' => 'User successfully removed'
+            ], Response::HTTP_OK);
+        } else {
+            return response()->json([
+                'message' => 'Action denied. Cannot delete yourself.'
+            ], Response::HTTP_FORBIDDEN);
+        }
     }
 }
